cleaning up package init - use separate namespaces
