+++ icd_api/__init__.py
@@ -1,4 +1,8 @@
-from .icd_api import *
-from .linearization import *
-from .util import *
-from .icd_entity import *
+# flake8: noqa
+
+from icd_api import util
+from icd_api import icd_util
+from icd_api import search_result
+from icd_api import linearization
+from icd_api import icd_entity
+from icd_api import icd_api
